extracting authorization code from fitbit redirect URL
