+++ kadence/pages/api/fitbit/getTokens.js
@@ -16,7 +16,7 @@ handler.post(async (req, res) => {
     const doc = {
         client_id: process.env.FITBIT_CLIENT_ID,
         grant_type: 'authorization_code',
-        redirect_uri: 'http://localhost:3000/fitbit', // CHANGE THIS TO A NEW URL
+        redirect_uri: 'http://localhost:3000/fitbit', //TODO CHANGE THIS TO A NEW URL
         code: req.body.authorizationCode,
         // code_verifier: "3q4u0f3f2830404x2j5d70483n6f200n593s2h261c0401076e0r6a5x0a6u3e313c516p55500u4162541b0t0f735u5d5b263k6n1w0e24504n6s15060e5i0w356n",
     };
