+++ kadence/pages/fitbit/index.js
@@ -5,7 +5,7 @@ import Button from '@mui/material/Button';
 import Textbox from '@/components/Textbox';
 import NetworkAPI from '@/lib/networkAPI';
 import PageLayout from '@/components/PageLayout';
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 import { Inter } from '@next/font/google';
 import { createTheme, ThemeProvider } from '@mui/material/styles';
 import { Dialog } from '@capacitor/dialog';
@@ -26,34 +26,21 @@ export default function Display() {
     async function connectDevice(e) {
         e.preventDefault();
         try {
-            const userData = {
+            /*const userData = {
                 username: localStorage.getItem('username'),
                 deviceName,
             };
-            await NetworkAPI.patch('/api/users/update', userData);
+            await NetworkAPI.patch('/api/users/update', userData);*/ //TODO Add this back in after testing
             const redirectUri =
                 process.env.NODE_ENV === 'development'
-                    ? 'http://localhost:3000/profile'
+                    ? 'http://localhost:3000/profile' 
                     : 'http://kadenceapp.com/profile';
             console.log(redirectUri);
+            localStorage.setItem('fromFitbit', 'true');
             window.location.assign(
                 `https://www.fitbit.com/oauth2/authorize?response_type=code&client_id=23QTD8&scope=activity+cardio_fitness+electrocardiogram+heartrate+location+nutrition+oxygen_saturation+profile+respiratory_rate+settings+sleep+social+temperature+weight&code_challenge=vaC5salqWAhM5k50MMvXGPxkTQGyQeLa0NpP_K3689Y&code_challenge_method=S256&state=3j3k386j3x606u7000324b4x4n0b0o06&redirect_uri=${encodeURI(
                     redirectUri
-                )}`
-            );
-            console.log(window.location.search);
-            const url = new URLSearchParams(window.location.search);
-            const authorizationCode = url.get('code');
-
-            const response = NetworkAPI.post('/api/fitbit/getTokens', {
-                authorizationCode,
-            });
-
-            localStorage.setItem('authorization_code', authorizationCode);
-            localStorage.setItem('access_token', response.json().access_token);
-            localStorage.setItem(
-                'refresh_token',
-                response.json().refresh_token
+                )}` //TODO Change this URL to only include the scope items we need (activity, cardio_fitness, electrocardiogram, heartrate, profile, settings)
             );
         } catch (err) {
             Dialog.alert({
